% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{friends}
\alias{friends}
\title{The transcript of Friends}
\format{
A \link[tibble:tibble-package]{tibble} with 67359 rows and
6)` variables:
\code{text}, \code{speaker}, \code{season}, \code{episode},
\code{scene} and \code{utterance}.
}
\source{
\url{https://github.com/emorynlp/character-mining}
}
\usage{
friends
}
\description{
Each season consists of episodes, each episode is divided into scenes,
each scene comprises utterances. One utterance per row.
Emotion annotation is included when available.
}
\keyword{datasets}
