% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{friends_emotions}
\alias{friends_emotions}
\title{Emotions for transcript of Friends}
\format{
A \link[tibble:tibble-package]{tibble} with 12606
rows an5)` variables:
\code{season}, \code{episode}, \code{scene}, \code{utterance} and
\code{entities}.
}
\source{
\url{https://github.com/emorynlp/character-mining}

\url{https://github.com/emorynlp/emotion-detection}
}
\usage{
friends_emotions
}
\description{
This \link[tibble:tibble-package]{tibble} contains the emotions for the
utterances where it is available.
}
\keyword{datasets}
