% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{friends_entities}
\alias{friends_entities}
\title{Character Entities for transcript of Friends}
\format{
A \link[tibble:tibble-package]{tibble} with 10557
rows an5)` variables:
\code{season}, \code{episode}, \code{scene}, \code{utterance} and
\code{entities}.
}
\source{
\url{https://github.com/emorynlp/character-mining}
}
\usage{
friends_entities
}
\description{
This \link[tibble:tibble-package]{tibble} contains the character entities for the
utterances where it is available.
}
\keyword{datasets}
