% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{friends_info}
\alias{friends_info}
\title{Episode Information}
\format{
A \link[tibble:tibble-package]{tibble} with 236 rows
an8)` variables:
\code{season}, \code{episode}, \code{title}, \code{directed_by},
\code{written_by}, \code{air_date}, \code{us_views_millions} and
\code{imdb_rating}.
}
\source{
\url{https://en.wikipedia.org/wiki/Friends}

\url{https://www.imdb.com/title/tt0108778/}
}
\usage{
friends_info
}
\description{
This \link[tibble:tibble-package]{tibble} contains additional information about
each of the episodes. Information was sourced from Wikipedia and IMDb on
August 26th, 2020.
}
\keyword{datasets}
